Move BranchPostTask to separate file
