+++ Branch-SDK/src/main/java/io/branch/referral/Branch.java
@@ -14,18 +14,15 @@ import android.content.pm.ActivityInfo;
 import android.content.pm.ApplicationInfo;
 import android.content.pm.PackageInfo;
 import android.content.pm.PackageManager;
-import android.graphics.drawable.Drawable;
 import android.net.Uri;
 import android.os.AsyncTask;
 import android.os.Bundle;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
-import androidx.annotation.StyleRes;
 
 import android.os.Handler;
 import android.os.Looper;
 import android.text.TextUtils;
-import android.view.View;
 
 import androidx.core.app.ActivityCompat;
 import io.branch.referral.Defines.PreinstallKey;
@@ -38,7 +35,6 @@ import java.io.UnsupportedEncodingException;
 import java.lang.ref.WeakReference;
 import java.net.HttpURLConnection;
 import java.net.URLEncoder;
-import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
@@ -212,7 +208,7 @@ public class Branch implements BranchViewHandler.IBranchViewEvents, SystemObserv
      * <p>A {@link Branch} object that is instantiated on init and holds the singleton instance of
      * the class during application runtime.</p>
      */
-    private static Branch branchReferral_;
+    public static Branch branchReferral_;
 
     private BranchRemoteInterface branchRemoteInterface_;
     final PrefHelper prefHelper_;
@@ -1562,7 +1558,7 @@ public class Branch implements BranchViewHandler.IBranchViewEvents, SystemObserv
 
     private void executeTimedBranchPostTask(final ServerRequest req, final int timeout) {
         final CountDownLatch latch = new CountDownLatch(1);
-        final BranchPostTask postTask = new BranchPostTask(req, latch);
+        final BranchPostTask postTask = new BranchPostTask(this, req, latch);
 
         postTask.executeTask();
         if (Looper.myLooper() == Looper.getMainLooper()) {
@@ -2148,13 +2144,6 @@ public class Branch implements BranchViewHandler.IBranchViewEvents, SystemObserv
     }
     
     /**
-     * <p>enum containing the sort options for return of credit history.</p>
-     */
-    public enum CreditHistoryOrder {
-        kMostRecentFirst, kLeastRecentFirst
-    }
-    
-    /**
      * Async Task to create  a short link for synchronous methods
      */
     private class GetShortLinkTask extends AsyncTask<ServerRequest, Void, ServerResponse> {
@@ -2162,197 +2151,6 @@ public class Branch implements BranchViewHandler.IBranchViewEvents, SystemObserv
             return branchRemoteInterface_.make_restful_post(serverRequests[0].getPost(),
                     prefHelper_.getAPIBaseUrl() + Defines.RequestPath.GetURL.getPath(),
                     Defines.RequestPath.GetURL.getPath(), prefHelper_.getBranchKey());
-        }
-    }
-
-    /**
-     * Asynchronous task handling execution of server requests. Execute the network task on background
-     * thread and request are  executed in sequential manner. Handles the request execution in
-     * Synchronous-Asynchronous pattern. Should be invoked only form main thread and  the results are
-     * published in the main thread.
-     */
-    private class BranchPostTask extends BranchAsyncTask<Void, Void, ServerResponse> {
-        ServerRequest thisReq_;
-        final CountDownLatch latch_;
-
-        public BranchPostTask(ServerRequest request, CountDownLatch latch) {
-            super();
-            thisReq_ = request;
-            latch_ = latch;
-        }
-
-        @Override
-        protected void onPreExecute() {
-            super.onPreExecute();
-            thisReq_.onPreExecute();
-            thisReq_.doFinalUpdateOnMainThread();
-        }
-
-        @Override
-        protected ServerResponse doInBackground(Void... voids) {
-            // update queue wait time
-            addExtraInstrumentationData(thisReq_.getRequestPath() + "-" + Defines.Jsonkey.Queue_Wait_Time.getKey(), String.valueOf(thisReq_.getQueueWaitTime()));
-            thisReq_.doFinalUpdateOnBackgroundThread();
-            if (isTrackingDisabled() && !thisReq_.prepareExecuteWithoutTracking()) {
-                return new ServerResponse(thisReq_.getRequestPath(), BranchError.ERR_BRANCH_TRACKING_DISABLED, "");
-            }
-            String branchKey = prefHelper_.getBranchKey();
-            ServerResponse result;
-            if (thisReq_.isGetRequest()) {
-                result = getBranchRemoteInterface().make_restful_get(thisReq_.getRequestUrl(), thisReq_.getGetParams(), thisReq_.getRequestPath(), branchKey);
-            } else {
-                result = getBranchRemoteInterface().make_restful_post(thisReq_.getPostWithInstrumentationValues(instrumentationExtraData_), thisReq_.getRequestUrl(), thisReq_.getRequestPath(), branchKey);
-            }
-            if (latch_ != null) {
-                latch_.countDown();
-            }
-            return result;
-        }
-
-        @Override
-        protected void onPostExecute(ServerResponse serverResponse) {
-            super.onPostExecute(serverResponse);
-            onPostExecuteInner(serverResponse);
-        }
-
-        void onPostExecuteInner(ServerResponse serverResponse) {
-            if (latch_ != null) {
-                latch_.countDown();
-            }
-            if (serverResponse == null) {
-                thisReq_.handleFailure(BranchError.ERR_BRANCH_INVALID_REQUEST, "Null response.");
-                return;
-            }
-
-            int status = serverResponse.getStatusCode();
-            if (status == 200) {
-                onRequestSuccess(serverResponse);
-            } else {
-                onRequestFailed(serverResponse, status);
-            }
-            networkCount_ = 0;
-
-            // In rare cases where this method is called directly (eg. when network calls time out),
-            // starting the next queue item can lead to stack over flow. Ensuring that this is
-            // queued back to the main thread mitigates this.
-            Handler handler = new Handler(Looper.getMainLooper());
-            handler.post(new Runnable() {
-                @Override
-                public void run() {
-                    processNextQueueItem();
-                }
-            });
-        }
-
-        private void onRequestSuccess(ServerResponse serverResponse) {
-            // If the request succeeded
-            @Nullable final JSONObject respJson = serverResponse.getObject();
-            if (respJson == null) {
-                thisReq_.handleFailure(500, "Null response json.");
-            }
-
-            if (thisReq_ instanceof ServerRequestCreateUrl && respJson != null) {
-                try {
-                    // cache the link
-                    BranchLinkData postBody = ((ServerRequestCreateUrl) thisReq_).getLinkPost();
-                    final String url = respJson.getString("url");
-                    linkCache_.put(postBody, url);
-                } catch (JSONException ex) {
-                    ex.printStackTrace();
-                }
-            } else if (thisReq_ instanceof ServerRequestLogout) {
-                //On Logout clear the link cache and all pending requests
-                linkCache_.clear();
-                requestQueue_.clear();
-            }
-
-
-            if (thisReq_ instanceof ServerRequestInitSession || thisReq_ instanceof ServerRequestIdentifyUserRequest) {
-                // If this request changes a session update the session-id to queued requests.
-                boolean updateRequestsInQueue = false;
-                if (!isTrackingDisabled() && respJson != null) {
-                    // Update PII data only if tracking is disabled
-                    try {
-                        if (respJson.has(Defines.Jsonkey.SessionID.getKey())) {
-                            prefHelper_.setSessionID(respJson.getString(Defines.Jsonkey.SessionID.getKey()));
-                            updateRequestsInQueue = true;
-                        }
-                        if (respJson.has(Defines.Jsonkey.RandomizedBundleToken.getKey())) {
-                            String new_Randomized_Bundle_Token = respJson.getString(Defines.Jsonkey.RandomizedBundleToken.getKey());
-                            if (!prefHelper_.getRandomizedBundleToken().equals(new_Randomized_Bundle_Token)) {
-                                //On setting a new Randomized Bundle Token clear the link cache
-                                linkCache_.clear();
-                                prefHelper_.setRandomizedBundleToken(new_Randomized_Bundle_Token);
-                                updateRequestsInQueue = true;
-                            }
-                        }
-                        if (respJson.has(Defines.Jsonkey.RandomizedDeviceToken.getKey())) {
-                            prefHelper_.setRandomizedDeviceToken(respJson.getString(Defines.Jsonkey.RandomizedDeviceToken.getKey()));
-                            updateRequestsInQueue = true;
-                        }
-                        if (updateRequestsInQueue) {
-                            updateAllRequestsInQueue();
-                        }
-                    } catch (JSONException ex) {
-                        ex.printStackTrace();
-                    }
-                }
-
-                if (thisReq_ instanceof ServerRequestInitSession) {
-                    setInitState(Branch.SESSION_STATE.INITIALISED);
-                    if (!((ServerRequestInitSession) thisReq_).handleBranchViewIfAvailable((serverResponse))) {
-                        checkForAutoDeepLinkConfiguration();
-                    }
-                    // Count down the latch holding getLatestReferringParamsSync
-                    if (getLatestReferringParamsLatch != null) {
-                        getLatestReferringParamsLatch.countDown();
-                    }
-                    // Count down the latch holding getFirstReferringParamsSync
-                    if (getFirstReferringParamsLatch != null) {
-                        getFirstReferringParamsLatch.countDown();
-                    }
-                }
-            }
-
-            if (respJson != null) {
-                thisReq_.onRequestSucceeded(serverResponse, branchReferral_);
-                requestQueue_.remove(thisReq_);
-            } else if (thisReq_.shouldRetryOnFail()) {
-                // already called handleFailure above
-                thisReq_.clearCallbacks();
-            } else {
-                requestQueue_.remove(thisReq_);
-            }
-        }
-
-        void onRequestFailed(ServerResponse serverResponse, int status) {
-            // If failed request is an initialisation request (but not in the intra-app linking scenario) then mark session as not initialised
-            if (thisReq_ instanceof ServerRequestInitSession && PrefHelper.NO_STRING_VALUE.equals(prefHelper_.getSessionParams())) {
-                setInitState(Branch.SESSION_STATE.UNINITIALISED);
-            }
-
-            // On a bad request or in case of a conflict notify with call back and remove the request.
-            if ((status == 400 || status == 409) && thisReq_ instanceof ServerRequestCreateUrl) {
-                ((ServerRequestCreateUrl) thisReq_).handleDuplicateURLError();
-            } else {
-                //On Network error or Branch is down fail all the pending requests in the queue except
-                //for request which need to be replayed on failure.
-                networkCount_ = 0;
-                thisReq_.handleFailure(status, serverResponse.getFailReason());
-            }
-
-            boolean unretryableErrorCode = (400 <= status && status <= 451) || status == BranchError.ERR_BRANCH_TRACKING_DISABLED;
-            // If it has an un-retryable error code, or it should not retry on fail, or the current retry count exceeds the max
-            // remove it from the queue
-            if (unretryableErrorCode || !thisReq_.shouldRetryOnFail() || (thisReq_.currentRetryCount >= prefHelper_.getNoConnectionRetryMax())) {
-                requestQueue_.remove(thisReq_);
-            } else {
-                // failure has already been handled
-                // todo does it make sense to retry the request without a callback? (e.g. CPID, LATD)
-                thisReq_.clearCallbacks();
-            }
-
-            thisReq_.currentRetryCount++;
         }
     }
